Remove v1 from API and package
